sspytorch/terri_utils.py
from torchvision import transforms
import numpy as np
import torch
def round2nearest_multiple(x, p):
    return ((x - 1) // p + 1) * p

normalize = transforms.Normalize(
            mean=[0.485, 0.456, 0.406],
            std=[0.229, 0.224, 0.225])


def img_transform(img):
    # 0-255 to 0-1
    img = np.float32(np.array(img)) / 255.
    img = img.transpose((2, 0, 1))
    img = normalize(torch.from_numpy(img.copy()))
    return img