sspytorch/requirements.txt
numpy
scipy
pytorch=0.4.1
torchvision
opencv3
yacs
tqdm
